Refactor schedule editing page to improve structure and add functionality
Refactors SchedulerEditPage.tsx, removes unused imports/components, and introduces tab-based layout.
